Expose la collectivité associée à l'aide dans le paquet NPM
