+++ package-aides-velo/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "aides-velo",
-	"version": "1.0.2",
+	"version": "1.1.1",
 	"description": "Les aides vélos proposées en France",
 	"author": "Maxime Quandalle <maxime@mesaidesvelo.fr>",
 	"license": "MIT",
@@ -28,6 +28,10 @@
 	},
 	"homepage": "https://github.com/mquandalle/mesaidesvelo/package-aides-velo#readme",
 	"devDependencies": {
+		"@etalab/decoupage-administratif": "^0.9.0-0",
+		"@rollup/plugin-alias": "^3.1.8",
+		"@rollup/plugin-json": "^4.1.0",
+		"@rollup/plugin-node-resolve": "^13.1.1",
 		"@rollup/plugin-typescript": "^8.3.0",
 		"@rollup/plugin-yaml": "^3.1.0",
 		"ava": "^3.15.0",
