+++ package-aides-velo/tsconfig.json
@@ -4,10 +4,15 @@
 		"moduleResolution": "node",
 		"declaration": true,
 		"esModuleInterop": true,
+		"allowJs": true,
+		"resolveJsonModule": true,
 		"outDir": "./build",
 		"forceConsistentCasingInFileNames": true,
 		"skipLibCheck": true,
-		"rootDir": "./"
+		"rootDir": "../",
+		"paths": {
+			"$lib/*": ["../src/lib/*"]
+		}
 	},
 	"include": ["global.d.ts", "index.ts"]
 }
