+++ src/routes/sitemap.txt.js
@@ -1,8 +1,6 @@
 // Un sitemap avec les 50 villes les plus peuplées
-import { loadJsonFile } from '$lib/readWriteJson';
 import { slugify } from '$lib/utils';
-
-const communes = loadJsonFile('src/data/communes.json');
+import communes from '$lib/data/communes.json';
 
 /** @type {import('@sveltejs/kit').RequestHandler} */
 export async function get({}) {
